package/python-cryptography: add host variant
Since commit 857c619cd76a6d83c15fe192a6aa70ad7d80673c,
python-cryptography depends on python3. So python3-cryptography became
redundant at that point. As python3-cryptography is a host-only package,
while python-cryptography is a python-only package, add a host variant
to python-cryptography to be able to drop python3-cryptography.

Update ca-certificates to use host-python-cryptography instead of
host-python3-cryptography.

While at it, drop host-python3-pip and host-python3-six as these
dependencies are not needed. Indeed, pip was never a dependency of
cryptography and six is not a dependency with python3 and will be
dropped in version 3.4 and
https://github.com/pyca/cryptography/commit/e66db8079d3fbd0110e87ece1fd48f4bfd9e48b9

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>

squash! package/python-cryptography: add host variant