package/openjdk: add ARC HS3x/4x support via "Zero Assembly Project"
This allows building and running a full-scale JVM in purely
interpretive mode on ARCv2 processors.

Once JIT'ed version is available for ARC we'll obviously switch to it
to gain a faster execution.

This is only supported for OpenJDK17, so we make OpenJDK11 unavailable
on ARC. However, there was a dependency difference between OpenJDK17
and OpenJDK11 (on host gcc >= 4.9). To avoid any potential
complexities if a package ever needs to "select BR2_PACKAGE_OPENJDK",
this commit moves this host gcc >= 4.9 dependency to
BR2_PACKAGE_OPENJDK itself.

Signed-off-by: Alexey Brodkin <abrodkin@synopsys.com>
Cc: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Cc: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Adam Duskett <aduskett@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>