package/rpi-firmware: update vcdbg runtime dependencies
We used to select the rpi-userland to make the vcdbg tool work but it
conflicts with the use of mesa. It implies we can not install vcdbg with
mesa libraries but it is false. We could use the prebuilt libraries from
the rpi-firmware sources to run vcdbg.

This patch removes the select dependency of rpi-userland package and
installs the prebuilt libraries needed by vcdbg only if rpi-userland
is not selected.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>
Reviewed-by: Bryce Schober <bryce.schober@gmail.com>
Tested-by: Bryce Schober <bryce.schober@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>