package/cog: bump to verion 0.12.0
This new release includes a number of improvements. Release notes:

  https://wpewebkit.org/release/cog-0.12.0.html

Small tweaks were needed due to the "fdo" platform backend being renamed
as "wl"; the symbol is kept as BR2_PACKAGE_COG_PLATFORM_FDO to avoid
breaking existing configurations. Also the new version supports building
with libsoup3, which is not yet packaged in Buildroot, so -DUSE_SOUP2=ON
is always passed to CMake.

Signed-off-by: Adrian Perez de Castro <aperez@igalia.com>
Tested-by: James Hilliard <james.hilliard1@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>