package/watchdogd: bump to version 3.5
This addresses the autobuilder failure for master as per 2021-11-27.
The root cause was a name clash in a dependency, the libite library,
clashing with the DirectFB LiTE library header files.  Hence, this
update alone does not fix [1], libite also needs to be updated.  To
provide a smooth transition though, watchdogd should be updated first.

[1]: http://autobuild.buildroot.net/results/185c753af2aa159b494b13f78b0826dddbe4aed6

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>