package/postgis: re-enable on Microblaze architecture
While adding this package with commit[1] Microblaze architecture has been
disabled by default pointing that it suffers from gcc ICE, but this is not
true, indeed the build error is referred to libgeos that is postgis
dependency:
```
  during RTL pass: reload
  .../bootlin-microblazeel-uclibc/build/libgeos-3.9.0/src/geom/util/Densifier.cpp: In static member function ‘static std::unique_ptr<std::vector<geos::geom::Coordinate> > geos::geom::util::Densifier::densifyPoints(geos::geom::Coordinate::Vect, double, const geos::geom::PrecisionModel*)’:
  .../bootlin-microblazeel-uclibc/build/libgeos-3.9.0/src/geom/util/Densifier.cpp:128:1: internal compiler error: in gen_reg_rtx, at emit-rtl.c:1155
  128 | }
      | ^
```

This build error has been worked around with this commit[2] and I've just
tested this build with Microblaze gcc 9.3/9.4/10.3 and 11.2 and it builds
succesfully since the bug above is gcc bug 90620[3] and it's been handled
in libgeos already.

So let's allow Microblaze(el/be) to build this postgis.

[1]: https://git.buildroot.net/buildroot/commit/?id=6e5cd44ca4e8d0fffcbb02e7f7ef8efd1b021c7f
[2]: https://git.buildroot.net/buildroot/commit/?id=ff206e7d8ff2e044e1e0391b4519335946400410
[3]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=90620

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>
Reviewed-by: Maxim Kochetkov <fido_max@inbox.ru>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>