package/networkd-dispatcher: new package
Networkd-dispatcher is a dispatcher daemon for systemd-networkd
connection status changes. This daemon is similar to
NetworkManager-dispatcher, but is much more limited in the types of
events it supports due to the limited nature of systemd-networkd.

To simplify the large number of transitive dependencies, remove the ones
that are implied by glibc.

To simplify the comment, simply don't show it if python is selected.
Python 2 is going to be removed soon anyway.

Signed-off-by: Michael Nosthoff <buildroot@heine.tech>
[Arnout: add Config.in comment and rework/simplify dependencies]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>